feat/#28: StoneFloating 컴포넌트로 분리
